Add files via upload
This was a project for my People Analytics class. The hypothetical client was an NBA team where I had to create a player valuation model to help assess three potential free agent candidates and compare career statistics between the three candidates. I created a presentation based on my analysis in which I recommended that the NBA team select one candidate for a free agent position and another candidate as a backup for another position.