Add files via upload
This was a class project for my People Analytics class where I created a regression model for the hypothetical client, Teachers for America. It identified the most relevant traits for a potential applicant completing their application. This would be useful for marketing to certain demographics of people in order to gain the most-desired applicants.